Covid 19 Detection using CT scans/app/requirements.txt
Flask==2.3.2
numpy
pillow
keras
tensorflow
gdown
